Add UnsafeUtils.DisplayException.
Sigh, another instance of TMIH (too much information hiding.)
There is a need to call either RTxFrame.PutExcept or
RTxStack.PutExcept, depending on the host exception
mechanism.  They have the same signature, and just
adding it to RTException.i3 would be what is needed.

But changing the RTS at the current time is a tangent I
don't want to go off on.  So I copied PutExcept into
UnsafeUtils.  This needs to be fixed properly.

Or, maybe, a version that sends the output to a TEXT or
Wr.T would be welcome.  Can it be done without RTS
changes?