1st commit post rename AssertionFailure->Backout.
Massive rename, but not everywhere.  (Assertions.)AssertionFailure
is separated into itself and (Failures.)Backout, as part of rework
to catch runtime errors, in addition to assertion failures, and
give the user the chance to back out of the failing operation.
Some places will still want to use AssertionFailure.  The
mechanism is not yet complete.

It compiles.  It bootstraps.  Does it do anything?  The world waits
in suspense.