Functioning rework of assertion+runtime failure handling.
Looks to be working as desired.  Still could use some cleanup of
leftover cruft, etc.

Catches and handles runtime errors and assertion failures coded as
calls on procedures in Assertions.i3.  If such an assertion failure is
not caught, it is changed to either:

  RuntimeError.E(UnhandledException), or
  RuntimeError.E(BlockedException),

as appropriate and reraised.  If either of these (occurring either
directly, or as converted above) is not caught, then a user dialog is
shown, giving the user options to termiate Schutz, Raise
Failures.Backout, or ignore the failure.

Failures.Backout is intended for editor functions to catch and back
out to a point where they can recover into a consistent state.  This
would normally include user-requested editor operations, backing out
to the state at the start of the operation.  It is unclear whether
this mechanism will work for things done on other than the worker
thread.

The ignore option is offered to the user only in case the original
failure started as Assertions.AssertionFailure, by using an internal
exception to just return from the Assert procedure without action.  It
is not clear whether compiler-generated raises would be translated by
the compiler so that ignoring would work.

The user dialog always writes a checkpoint of the open image, if there
is one.  It is a GUI dialog, if the the GUI editor is running, else it
is a command-line dialog on stderr and stdin.  Even in the GUI case,
there is a record of what happened written to stderr.

<*ASSERT*> failures are runtime exceptions, so get handled as such.