Background work for diagnosing colors/fonts bitrot.
Add a bunch of locking and rework LL pragrams, primarily
regarding VBT.mu and v : VBT.T.

Various experimental things regarding PaintText sometimes
not painting.  The first call on VBT.PaintText works as
expected and so does every subsequent call that passes the same
font and paint op as the first.  But any that passes either
a different font or different paint op does not paint
anything.  In the end, add global Options.OneFont and
Options.DullColors to cause all PaintText calls to use
the same values, as a way to make Schutz useable, 'tho
not as attractive of helpful with fonts and colors.

Move some derived info so as to not to be unnecessarily
recomputed.