Fix a bug miscomputing NewLinesCt in MergeTextEdit.
This was being incremented in two different ways, the wrong way being
when a new mark was built.  That is now changed to a new variable
MteMarkCt, which is not used, but left there for possible help in
future debugging.

This fixes a bug shown by test14_152, which has not runtime failures,
but still has an output mismatch.