Add many cases where TkmEstRef needs to be patched later.
The needed value is not available when the mark is initially
constructed.  There was already a mechanism for patching the
EstNodeCt field, for the same reason.

This is only for internal info, help debugging, etc.  It is
not expected to fix test failures other than an assertion
failure checking that the value is there.