Incomplete rework of TextEdit.m3, for adjusting marks.
Including some temporary stuff for debugging.

This exercise led to the belief that the existing method of
adjusting marks after MergeTextEdit, which never has been
right, is not a workable approach.  I plan a different way,
driven mostly by the linked lists of marked points and linesref
lists, rather than primarily by changes in the Est.